Implemented access control. Guests can only view posts and Authenticated users can only perform edit and delete operations on posts belonging to them.

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use Illuminate\Support\Facades\Auth;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -73,6 +83,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(Auth::id() !== $post->user_id) {
+            return redirect("/posts")->with('error', 'Unauthorized Page');
+        }
+
         return view("posts.edit")->with("post", $post);
     }
 
@@ -85,13 +101,19 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $post = Post::find($id);
+
+        // Check for correct user
+        if(Auth::id() !== $post->user_id) {
+            return redirect("/posts")->with('error', 'Unauthorized Page');
+        }
+
         $validated = $request->validate([
             'title' => 'required',
             'body' => 'required'
         ]);
 
         // Update Post
-        $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->save();
@@ -108,6 +130,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(Auth::id() !== $post->user_id) {
+            return redirect("/posts")->with('error', 'Unauthorized Page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
